run_gromacs.sh: check if file for presimulation exist

+++ share/scripts/inverse/run_gromacs.sh
@@ -78,6 +78,7 @@ if [[ $(csg_get_property cg.inverse.gromacs.pre_simulation) = "yes" && $1 != "--
     for i in mdp topol_in index; do
       f="$(csg_get_property "cg.inverse.gromacs.pre_simulation.$i" "${!i}")" #filter me away
       [[ $f != ${!i} ]] && ((cp++))
+      [[ -f ${f} ]] || die "${0##*/}: file '$f' not found (make sure it is in cg.inverse.filelist)"
       critical cp "${f}" "./pre_simulation/${!i}"
     done
     [[ $cp -eq 0 ]] && die "${0##*/}: mdp, topol_in and index of the presimulation are the same as for the main simulation, that does not make sense - at least one has to be different!"
